split mockHttpCalls into separate files
