+++ testUtils/mockHttpCalls/auth.js
@@ -1,0 +1,3 @@
+export const authTokenResponse = {
+  auth_token: 'efeb3f94-8deb-41a0-84cc-713801ff165e',
+};
