+++ testUtils/mockHttpCalls/status.js
@@ -1,0 +1,349 @@
+import { Providers } from 'shared/constants';
+import { V4_CLUSTER } from './constantsAndHelpers';
+
+export const v4AWSClusterStatusResponse = {
+  aws: {
+    availabilityZones: [
+      {
+        name: 'eu-central-1a',
+        subnet: {
+          private: { cidr: '10.1.2.0/25' },
+          public: { cidr: '10.1.2.128/25' },
+        },
+      },
+    ],
+    autoScalingGroup: { name: '' },
+  },
+  cluster: {
+    conditions: [
+      {
+        lastTransitionTime: '2019-11-15T15:54:05.711696992Z',
+        status: 'True',
+        type: 'Creating',
+      },
+    ],
+    network: { cidr: '10.1.2.0/24' },
+    nodes: [
+      {
+        labels: {
+          'aws-operator.giantswarm.io/version': '5.5.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AWSInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'eu-central-1',
+          'failure-domain.beta.kubernetes.io/zone': 'eu-central-1c',
+          'giantswarm.io/provider': Providers.AWS,
+          ip: '10.1.2.18',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname':
+            'ip-10-1-2-18.eu-central-1.compute.internal',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-14T05:28:11.958410127Z',
+        name: 'ip-10-1-2-18.eu-central-1.compute.internal',
+        version: '5.5.0',
+      },
+      {
+        labels: {
+          'aws-operator.giantswarm.io/version': '5.5.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AWSInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'eu-central-1',
+          'failure-domain.beta.kubernetes.io/zone': 'eu-central-1c',
+          'giantswarm.io/provider': Providers.AWS,
+          ip: '10.1.2.49',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname':
+            'ip-10-1-2-49.eu-central-1.compute.internal',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'master',
+          'node-role.kubernetes.io/master': '',
+          'node.kubernetes.io/master': '',
+          role: 'master',
+        },
+        lastTransitionTime: '2019-11-14T05:28:11.958410751Z',
+        name: 'ip-10-1-2-49.eu-central-1.compute.internal',
+        version: '5.5.0',
+      },
+      {
+        labels: {
+          'aws-operator.giantswarm.io/version': '5.5.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AWSInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'eu-central-1',
+          'failure-domain.beta.kubernetes.io/zone': 'eu-central-1c',
+          'giantswarm.io/provider': Providers.AWS,
+          ip: '10.1.2.52',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname':
+            'ip-10-1-2-52.eu-central-1.compute.internal',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-14T05:28:11.958411157Z',
+        name: 'ip-10-1-2-52.eu-central-1.compute.internal',
+        version: '5.5.0',
+      },
+      {
+        labels: {
+          'aws-operator.giantswarm.io/version': '5.5.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AWSInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'eu-central-1',
+          'failure-domain.beta.kubernetes.io/zone': 'eu-central-1c',
+          'giantswarm.io/provider': Providers.AWS,
+          ip: '10.1.2.85',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname':
+            'ip-10-1-2-85.eu-central-1.compute.internal',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-14T05:28:11.958411586Z',
+        name: 'ip-10-1-2-85.eu-central-1.compute.internal',
+        version: '5.5.0',
+      },
+    ],
+    resources: null,
+    scaling: { desiredCapacity: 3 },
+    versions: [],
+  },
+};
+
+export const v4AzureClusterStatusResponse = {
+  aws: { availabilityZones: null, autoScalingGroup: { name: '' } },
+  cluster: {
+    conditions: [
+      {
+        lastTransitionTime: '2019-11-29T16:03:28.054594841Z',
+        status: 'True',
+        type: 'Created',
+      },
+    ],
+    network: { cidr: '' },
+    nodes: [
+      {
+        labels: {
+          'azure-operator.giantswarm.io/version': '2.7.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AzureInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'westeurope',
+          'failure-domain.beta.kubernetes.io/zone': '0',
+          'giantswarm.io/provider': Providers.AZURE,
+          ip: '10.1.0.5',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'o7oyb-master-000000',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'master',
+          'node-role.kubernetes.io/master': '',
+          'node.kubernetes.io/master': '',
+          role: 'master',
+        },
+        lastTransitionTime: '2019-11-29T16:03:27.992976986Z',
+        name: 'o7oyb-master-000000',
+        version: '2.7.0',
+      },
+      {
+        labels: {
+          'azure-operator.giantswarm.io/version': '2.7.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AzureInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'westeurope',
+          'failure-domain.beta.kubernetes.io/zone': '4',
+          'giantswarm.io/provider': Providers.AZURE,
+          ip: '10.1.1.8',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'o7oyb-worker-000004',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-29T16:03:27.992978386Z',
+        name: 'o7oyb-worker-000004',
+        version: '2.7.0',
+      },
+      {
+        labels: {
+          'azure-operator.giantswarm.io/version': '2.7.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AzureInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'westeurope',
+          'failure-domain.beta.kubernetes.io/zone': '0',
+          'giantswarm.io/provider': Providers.AZURE,
+          ip: '10.1.1.9',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'o7oyb-worker-000005',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-29T16:03:27.992978986Z',
+        name: 'o7oyb-worker-000005',
+        version: '2.7.0',
+      },
+      {
+        labels: {
+          'azure-operator.giantswarm.io/version': '2.7.0',
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/instance-type': V4_CLUSTER.AzureInstanceType,
+          'beta.kubernetes.io/os': 'linux',
+          'failure-domain.beta.kubernetes.io/region': 'westeurope',
+          'failure-domain.beta.kubernetes.io/zone': '1',
+          'giantswarm.io/provider': Providers.AZURE,
+          ip: '10.1.1.10',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'o7oyb-worker-000006',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-11-29T16:03:27.992979386Z',
+        name: 'o7oyb-worker-000006',
+        version: '2.7.0',
+      },
+    ],
+    resources: [
+      {
+        conditions: [
+          {
+            lastTransitionTime: '0001-01-01T00:00:00Z',
+            status: 'InstancesUpgrading',
+            type: 'Stage',
+          },
+        ],
+        name: 'instancev11',
+      },
+    ],
+    scaling: { desiredCapacity: 0 },
+    versions: [
+      {
+        date: '0001-01-01T00:00:00Z',
+        lastTransitionTime: '2019-11-29T16:03:28.105736089Z',
+        semver: '2.7.0',
+      },
+    ],
+  },
+};
+
+export const v4KVMClusterStatusResponse = {
+  aws: { availabilityZones: null, autoScalingGroup: { name: '' } },
+  cluster: {
+    conditions: [
+      {
+        lastTransitionTime: '2019-11-27T09:22:45.131944981Z',
+        status: 'True',
+        type: 'Created',
+      },
+    ],
+    network: { cidr: '' },
+    nodes: [
+      {
+        labels: {
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/os': 'linux',
+          'giantswarm.io/provider': Providers.KVM,
+          ip: '172.23.9.10',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'master-w0je0-7fcb8856b6-4r8jj',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'master',
+          'kvm-operator.giantswarm.io/version': '3.10.0',
+          'node-role.kubernetes.io/master': '',
+          'node.kubernetes.io/master': '',
+          role: 'master',
+        },
+        lastTransitionTime: '2019-12-02T04:36:46.868446664Z',
+        name: 'master-w0je0-7fcb8856b6-4r8jj',
+        version: '3.10.0',
+      },
+      {
+        labels: {
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/os': 'linux',
+          'giantswarm.io/provider': Providers.KVM,
+          ip: '172.23.9.194',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'worker-3c8x8-85c74757f5-x95d5',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'kvm-operator.giantswarm.io/version': '3.10.0',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-12-02T04:36:46.868447575Z',
+        name: 'worker-3c8x8-85c74757f5-x95d5',
+        version: '3.10.0',
+      },
+      {
+        labels: {
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/os': 'linux',
+          'giantswarm.io/provider': Providers.KVM,
+          ip: '172.23.9.186',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'worker-87jfz-bf8df564c-jnst9',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'kvm-operator.giantswarm.io/version': '3.10.0',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-12-02T04:36:46.868448272Z',
+        name: 'worker-87jfz-bf8df564c-jnst9',
+        version: '3.10.0',
+      },
+      {
+        labels: {
+          'beta.kubernetes.io/arch': 'amd64',
+          'beta.kubernetes.io/os': 'linux',
+          'giantswarm.io/provider': Providers.KVM,
+          ip: '172.23.9.234',
+          'kubernetes.io/arch': 'amd64',
+          'kubernetes.io/hostname': 'worker-r0e8m-696fd6c7bc-slgsh',
+          'kubernetes.io/os': 'linux',
+          'kubernetes.io/role': 'worker',
+          'kvm-operator.giantswarm.io/version': '3.10.0',
+          'node-role.kubernetes.io/worker': '',
+          'node.kubernetes.io/worker': '',
+          role: 'worker',
+        },
+        lastTransitionTime: '2019-12-02T04:36:46.868448804Z',
+        name: 'worker-r0e8m-696fd6c7bc-slgsh',
+        version: '3.10.0',
+      },
+    ],
+    resources: null,
+    scaling: { desiredCapacity: 0 },
+    versions: [
+      {
+        date: '0001-01-01T00:00:00Z',
+        lastTransitionTime: '2019-11-27T09:22:45.332250352Z',
+        semver: '3.10.0',
+      },
+    ],
+  },
+};
