+++ testUtils/mockHttpCalls/user.js
@@ -1,0 +1,7 @@
+import { USER_EMAIL } from './constantsAndHelpers';
+
+export const userResponse = {
+  email: USER_EMAIL,
+  created: '2019-09-19T12:40:16.2231629Z',
+  expiry: '2020-01-01T00:00:00Z',
+};
